requirements.txt
obs-websocket-py
python-dateutil
pyserial
pyaudio